refactor api docs (in progress)
