+++ app/forms.py
@@ -2,7 +2,8 @@ from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, BooleanField, SubmitField
 from wtforms.validators import ValidationError, DataRequired, \
     Email, EqualTo
-import models
+
+from app import models
 
 
 class LoginForm(FlaskForm):
