+++ app/templates/api_help.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <div id = "container">
+        <p> <h1 align = "center">Документация API</h1>
+        <p> <h4>тут про пост</h4>
+        <p> <h4>пример</h4>
+    </div>
+{% endblock %}
