+++ app/templates/chart.html
